Added new U correction functions, implemented Modified Gaussian, resolved all compilation warnings
git-svn-id: https://svs.icts.kuleuven.be/repo/svs_project058/branches/hayen@83 ad58af76-793d-47a9-990c-b06e191431a5

+++ trunk/NME/include/NuclearUtilities.h
@@ -11,7 +11,7 @@ namespace NuclearStructure {
  *
  * @param x double
  */
-inline double sign(double x) { return (x == 0) ? 0 : x / std::abs(x); }
+inline int sign(double x) { return (x == 0) ? 0 : (int)(x / std::abs(x)); }
 
 /**
  * The Kronecker delta for two doubles
